Support gremlin consoles in entrypoint
This allows for either of the following to be given as the docker command:

`gremlin-console`
Launches a gremlin console, directly attached to the storage layer of choice.

`gremlin-console-remote`
Launches a gremlin console which attaches to the remote ports that the gremlin
server is exposing.

+++ 0.3/docker-entrypoint.sh
@@ -22,7 +22,8 @@ JANUS_PROPS="${JANUS_CONFIG_DIR}/janusgraph.properties"
 GREMLIN_YAML="${JANUS_CONFIG_DIR}/gremlin-server.yaml"
 
 # running as root; step down to run as janusgraph user
-if [ "$1" == 'janusgraph' ] && [ "$(id -u)" == "0" ]; then
+if [[ "$(id -u)" == "0" ]] && [[ "$1" == 'janusgraph' || "$1" == 'gremlin-console' || "$1" == 'gremlin-console-remote' ]]; then
+  echo "Running root setup steps for '$1'"
   mkdir -p ${JANUS_DATA_DIR} ${JANUS_CONFIG_DIR}
   chown -R janusgraph:janusgraph ${JANUS_DATA_DIR} ${JANUS_CONFIG_DIR}
   usermod -d ${JANUS_HOME} janusgraph
@@ -32,7 +33,8 @@ if [ "$1" == 'janusgraph' ] && [ "$(id -u)" == "0" ]; then
 fi
 
 # running as non root user
-if [ "$1" == 'janusgraph' ]; then
+if [[ "$1" == 'janusgraph' || "$1" == 'gremlin-console' || "$1" == 'gremlin-console-remote' ]] ; then
+  echo "Running non-root setup steps for '$1'"
   # setup config directory
   mkdir -p ${JANUS_DATA_DIR} ${JANUS_CONFIG_DIR}
   cp conf/gremlin-server/janusgraph-${JANUS_PROPS_TEMPLATE}-server.properties ${JANUS_CONFIG_DIR}/janusgraph.properties
@@ -40,6 +42,11 @@ if [ "$1" == 'janusgraph' ]; then
   chown -R "$(id -u):$(id -g)" ${JANUS_DATA_DIR} ${JANUS_CONFIG_DIR}
   chmod 700 ${JANUS_DATA_DIR} ${JANUS_CONFIG_DIR}
   chmod -R 600 ${JANUS_CONFIG_DIR}/*
+
+  # override hosts for remote connections with Gremlin Console
+  if ! [ -z "${GREMLIN_REMOTE_HOSTS:-}" ]; then
+    sed -i "s/hosts\s*:.*/hosts: [$GREMLIN_REMOTE_HOSTS]/" ${JANUS_HOME}/conf/remote.yaml
+  fi
 
   # apply configuration from environment
   while IFS='=' read -r envvar_key envvar_val; do
@@ -69,22 +76,31 @@ if [ "$1" == 'janusgraph' ]; then
     cat "$GREMLIN_YAML"
     exit 0
   else
-    # wait for storage
-    if ! [ -z "${JANUS_STORAGE_TIMEOUT:-}" ]; then
-      F="$(mktemp --suffix .groovy)"
-      echo "graph = JanusGraphFactory.open('${JANUS_CONFIG_DIR}/janusgraph.properties')" > $F
-      timeout -k 10 "${JANUS_STORAGE_TIMEOUT}s" bash -c \
-        "until bin/gremlin.sh -e $F > /dev/null 2>&1; do echo \"waiting for storage...\"; sleep 5; done"
-      rm -f "$F"
+    if [ "$1" == 'janusgraph' ] ; then
+      # wait for storage
+      if ! [ -z "${JANUS_STORAGE_TIMEOUT:-}" ]; then
+        F="$(mktemp --suffix .groovy)"
+        echo "graph = JanusGraphFactory.open('${JANUS_CONFIG_DIR}/janusgraph.properties')" > $F
+        timeout -k 10 "${JANUS_STORAGE_TIMEOUT}s" bash -c \
+          "until bin/gremlin.sh -e $F > /dev/null 2>&1; do echo \"waiting for storage...\"; sleep 5; done"
+        rm -f "$F"
+      fi
+      echo "Starting janusgraph gremlin server"
+      exec ${JANUS_HOME}/bin/gremlin-server.sh ${JANUS_CONFIG_DIR}/gremlin-server.yaml
     fi
 
-    exec ${JANUS_HOME}/bin/gremlin-server.sh ${JANUS_CONFIG_DIR}/gremlin-server.yaml
-  fi
-fi
+    if [ "$1" == 'gremlin-console-remote' ] ; then
+      echo "Starting gremlin remote console"
+      exec ${JANUS_HOME}/bin/gremlin.sh -i scripts/remote-connect.groovy
+    fi
 
-# override hosts for remote connections with Gremlin Console
-if ! [ -z "${GREMLIN_REMOTE_HOSTS:-}" ]; then
-  sed -i "s/hosts\s*:.*/hosts: [$GREMLIN_REMOTE_HOSTS]/" ${JANUS_HOME}/conf/remote.yaml
+    if [ "$1" == 'gremlin-console' ] ; then
+      echo "Starting gremlin console"
+      C=$(mktemp --suffix .groovy)
+      echo "graph = JanusGraphFactory.open('${JANUS_CONFIG_DIR}/janusgraph.properties')" > $C
+      exec ${JANUS_HOME}/bin/gremlin.sh -i $C
+    fi
+  fi
 fi
 
 exec "$@"
